+++ doc/source/changelog.rst
@@ -52,6 +52,8 @@ organisation on `GitHub <https://github.com/openbiosim/sire>`__.
 
 * Print warning when ``sire.legacy.IO.AmberPrm`` parser re-orders residues due to covalent bonds between molecules.
 
+* Added internal implementation of virtual sites in the Sire-to-OpenMM conversion layer 
+
 `2025.4.0 <https://github.com/openbiosim/sire/compare/2025.3.0...2025.4.0>`__ - February 2026
 ---------------------------------------------------------------------------------------------